# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/mailing_subscriptions:
    get:
      tags:
        - MailingSubscriptions
      summary: List MailingSubscriptions
      description: Get a paginated list of MailingSubscriptions with optional filtering and sorting
      operationId: listMailingSubscriptions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingSubscriptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MailingSubscriptions
      summary: Create MailingSubscription
      description: Create a new MailingSubscription
      operationId: createMailingSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMailingSubscriptionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailingSubscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    MailingSubscription:
      type: object
      required:
        - id
        - contact_id
        - mailing_audience_id
        - opt_out
        - metadata
      properties:
        id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        mailing_audience_id:
          type: string
          format: uuid
        opt_out:
          type: boolean
        opt_out_datetime:
          nullable: true
          type: string
          format: date-time
        opt_out_reason_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateMailingSubscriptionRequest:
      type: object
      required:
        - contact_id
        - mailing_audience_id
        - opt_out
        - metadata
      properties:
        contact_id:
          type: string
          format: uuid
        mailing_audience_id:
          type: string
          format: uuid
        opt_out:
          type: boolean
        opt_out_datetime:
          nullable: true
          type: string
          format: date-time
        opt_out_reason_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    MailingSubscriptionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MailingSubscription'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []